+++ packages/climate-ref-esmvaltool/src/climate_ref_esmvaltool/diagnostics/__init__.py
@@ -11,7 +11,7 @@ from climate_ref_esmvaltool.diagnostics.cloud_scatterplots import (
 )
 from climate_ref_esmvaltool.diagnostics.ecs import EquilibriumClimateSensitivity
 from climate_ref_esmvaltool.diagnostics.example import GlobalMeanTimeseries
-from climate_ref_esmvaltool.diagnostics.sea_ice_area_seasonal_cycle import SeaIceAreaSeasonalCycle
+from climate_ref_esmvaltool.diagnostics.sea_ice_area_basic import SeaIceAreaBasic
 from climate_ref_esmvaltool.diagnostics.tcr import TransientClimateResponse
 from climate_ref_esmvaltool.diagnostics.tcre import TransientClimateResponseEmissions
 from climate_ref_esmvaltool.diagnostics.zec import ZeroEmissionCommitment
@@ -26,7 +26,7 @@ __all__ = [
     "CloudScatterplotsReference",
     "EquilibriumClimateSensitivity",
     "GlobalMeanTimeseries",
-    "SeaIceAreaSeasonalCycle",
+    "SeaIceAreaBasic",
     "TransientClimateResponse",
     "TransientClimateResponseEmissions",
     "ZeroEmissionCommitment",