+++ packages/climate-ref-esmvaltool/tests/unit/diagnostics/test_sea_ice_area.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 import pandas
-from climate_ref_esmvaltool.diagnostics import SeaIceAreaSeasonalCycle
+from climate_ref_esmvaltool.diagnostics import SeaIceAreaBasic
 from climate_ref_esmvaltool.recipe import load_recipe
 
 from climate_ref_core.datasets import SourceDatasetType
@@ -12,9 +12,9 @@ def test_update_recipe():
         SourceDatasetType.CMIP6: pandas.read_json(Path(__file__).parent / "input_files_sea_ice_area.json")
     }
     recipe = load_recipe("ref/recipe_ref_sea_ice_area_basic.yml")
-    SeaIceAreaSeasonalCycle().update_recipe(recipe, input_files)
+    SeaIceAreaBasic().update_recipe(recipe, input_files)
     assert len(recipe["datasets"]) == 1
     dataset = recipe["datasets"][0]
     assert dataset["dataset"] == input_files[SourceDatasetType.CMIP6].iloc[0].source_id
-    assert dataset["timerange"] == "1995/2014"
+    assert "timerange" not in dataset
     assert len(recipe["diagnostics"]) == 2