+++ packages/climate-ref-pmp/tests/integration/test_diagnostics.py
@@ -9,6 +9,9 @@ diagnostics = [pytest.param(diagnostic, id=diagnostic.slug) for diagnostic in pm
 @pytest.mark.slow
 @pytest.mark.parametrize("diagnostic", diagnostics)
 def test_diagnostics(diagnostic: Diagnostic, diagnostic_validation):
+    if "enso" in diagnostic.slug:
+        pytest.skip("ENSO diagnostics do not support the sample data")
+
     validator = diagnostic_validation(diagnostic)
 
     definition = validator.get_definition()
@@ -21,3 +24,4 @@ def test_build_results(diagnostic: Diagnostic, diagnostic_validation):
 
     definition = validator.get_regression_definition()
     validator.validate(definition)
+    validator.execution_regression.check(definition.key, definition.output_directory)