+++ packages/climate-ref/src/climate_ref/testing.py
@@ -13,20 +13,21 @@ from climate_ref.executor import handle_execution_result
 from climate_ref.models import Execution, ExecutionGroup
 from climate_ref_core.dataset_registry import dataset_registry_manager, fetch_all_files
 from climate_ref_core.diagnostics import Diagnostic, ExecutionResult
+from climate_ref_core.env import env
 from climate_ref_core.pycmec.metric import CMECMetric
 from climate_ref_core.pycmec.output import CMECOutput
 
 
 def _determine_test_directory() -> Path | None:
-    expected = Path(__file__).parents[4] / "tests" / "test-data"
+    path = env.path("REF_TEST_DATA_DIR", default=Path(__file__).parents[4] / "tests" / "test-data")
 
-    if not expected.exists():  # pragma: no cover
+    if not path.exists():  # pragma: no cover
         return None
-    return expected
+    return path
 
 
 TEST_DATA_DIR = _determine_test_directory()
-SAMPLE_DATA_VERSION = "v0.6.0"
+SAMPLE_DATA_VERSION = "v0.6.3"
 
 
 def fetch_sample_data(force_cleanup: bool = False, symlink: bool = False) -> None:
@@ -102,6 +103,7 @@ def validate_result(diagnostic: Diagnostic, config: Config, result: ExecutionRes
 
     # Validate bundles
     metric_bundle = CMECMetric.load_from_json(result.to_output_path(result.metric_bundle_filename))
+    CMECMetric.model_validate(metric_bundle)
     bundle_dimensions = tuple(metric_bundle.DIMENSIONS.root["json_structure"])
     assert diagnostic.facets == bundle_dimensions, bundle_dimensions
     CMECOutput.load_from_json(result.to_output_path(result.output_bundle_filename))